+++ projects/js-packages/publicize-components/src/components/admin-page/test/index.test.jsx
@@ -2,7 +2,7 @@ import { render, renderHook, screen } from '@testing-library/react';
 import { useSelect, createReduxStore, register } from '@wordpress/data';
 import React from 'react';
 import { SOCIAL_STORE_CONFIG, SOCIAL_STORE_ID } from '../../../social-store';
-import Admin from '../index';
+import { SocialAdminPage } from '../index';
 
 const store = createReduxStore( SOCIAL_STORE_ID, SOCIAL_STORE_CONFIG );
 register( store );
@@ -25,10 +25,10 @@ describe( 'load the app', () => {
 	test( 'container renders', () => {
 		let storeSelect;
 		renderHook( () => useSelect( select => ( storeSelect = select( SOCIAL_STORE_ID ) ) ) );
-		jest.spyOn( storeSelect, 'getSocialPluginSettings' ).mockReset().mockReturnValue( {
-			show_pricing_page: true,
+		jest.spyOn( storeSelect, 'getSocialSettings' ).mockReset().mockReturnValue( {
+			showPricingPage: true,
 		} );
-		render( <Admin /> );
+		render( <SocialAdminPage /> );
 		expect( screen.getByText( `Jetpack Social ${ version }` ) ).toBeInTheDocument();
 	} );
 } );