+++ projects/packages/publicize/src/class-publicize-setup.php
@@ -50,6 +50,8 @@ class Publicize_Setup {
 				new $controller();
 			}
 		}
+
+		Social_Admin_Page::init();
 	}
 
 	/**